Stop computing the default Cloud SDK location preference But do show the computed location in the Cloud SDK preference page.
