+++ plugins/com.google.cloud.tools.eclipse.sdk/plugin.xml
@@ -1,10 +1,4 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <?eclipse version="3.4"?>
 <plugin>
-   <extension
-         point="org.eclipse.equinox.preferences.preferences">
-      <initializer
-            class="com.google.cloud.tools.eclipse.sdk.internal.PreferenceInitializer">
-      </initializer>
-   </extension>
 </plugin>
